// src/common/gas-parts/types.ts
export enum FileType {
    SETUP = "MontessoriChacalaSchoolGroupSetup",
    REPORT = "MontessoriChacalaSchoolReport",
}

export enum Urls {
    MEDIA_SERVER = "https://media.githubusercontent.com/media/YamanquiChacala/GeneradorDeReportes/refs/heads/main/",
}

export enum Colors {
    LOGO_OSCURO = "#34666A",
    LOGO_CLARO = "#159A5E",
    ORANGE = "#EA4335",
    LANGUAGE = "#c9daf8",
    SCIENCE = "#fce5cd",
    NATURE = "#d9ead3",
    HUMANITIES = "#ead1dc",
}